Reload syntax highlighting with double <space>
Jumping between Vim being in the background and in the foreground and then
between various tabs (e.g. git commit message compilation tab) sometimes breaks
syntax highlighting. With this the quick shortcut can remedy this.

Don't know why it's <leader><space>, tbh. Not sure if it makes sense for it to
be explicitly space one time and leader the other. No reason to change at the
moment, though.